sql: use ClearRange for index truncation
This change makes the deletion of index data use the ClearRange batch
request. DROP INDEX schema changes in the same transaction as the one
which created the table still uses the slower DelRange because
ClearRange cannot be run inside a transaction and will remove write
intents of the index keys which have not been resolved in the
transaction.

This deletion of index data happens once the GC TTL period has passed
and it is safe to remove all the data. See PR #30566 for which delays
this data deletion.

The GC threshold is also forwarded for the cleared range to prevent
reads and writes of index data.

Fixes #20696.

Release note (performance improvement): Deletion of index data is faster.

+++ pkg/sql/backfill.go
@@ -255,14 +255,19 @@ func (sc *SchemaChanger) truncateIndexes(
 				if err := td.init(txn, nil /* *tree.EvalContext */); err != nil {
 					return err
 				}
-				resume, err = td.deleteIndex(
-					ctx,
-					&desc,
-					resumeAt,
-					chunkSize,
-					noAutoCommit,
-					false, /* traceKV */
-				)
+				if !sc.canClearRangeForDrop(&desc) {
+					resume, err = td.deleteIndex(
+						ctx,
+						&desc,
+						resumeAt,
+						chunkSize,
+						noAutoCommit,
+						false, /* traceKV */
+					)
+				} else {
+					err = td.clearIndex(ctx, &desc)
+					resume = roachpb.Span{}
+				}
 				done = resume.Key == nil
 				return err
 			}); err != nil {
